requirements.txt
flask
paho.mqtt